Fix a mutable-export warning by flattening connectors/mongo.js onto connectors.js
