+++ packages/vulcan-lib/lib/server/main.js
@@ -2,8 +2,6 @@
 import './intl_polyfill.js';
 import './site.js';
 
-import './connectors/mongo.js';
-
 export * from './connectors.js';
 export * from './query.js';
 export * from '../modules/index.js';
